src/services/deck_recommender/requirements.txt
fastapi
uvicorn
uvloop
orjson
pyyaml
zstandard
setproctitle
sekai-deck-recommend-cpp